apps/workflow-api/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic>=2.9,<3.0
httpx>=0.27,<1.0
PyYAML>=6.0,<7.0
